// frontend/src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
// ...
      schemes: {
        Row: {
          application_link: string | null
          benefits: string | null
          category: string | null
          created_at: string
          description: string
          eligibility: string | null
          id: string
          is_active: boolean
          ministry: string | null
          slug: string
          state: string | null
          title: string
          updated_at: string
        }
        Insert: {
          application_link?: string | null
          benefits?: string | null
          category?: string | null
          created_at?: string
          description: string
          eligibility?: string | null
          id?: string
          is_active?: boolean
          ministry?: string | null
          slug: string
          state?: string | null
          title: string
          updated_at?: string
        }
        Update: {
          application_link?: string | null
          benefits?: string | null
          category?: string | null
          created_at?: string
          description?: string
          eligibility?: string | null
          id?: string
          is_active?: boolean
          ministry?: string | null
          slug?: string
          state?: string | null
          title?: string
          updated_at?: string
        }
        Relationships: []
      }
    }
// ...
  }
}